move replay page to pages
