+++ src/renderer/pages/replays/create.tsx
@@ -1,0 +1,7 @@
+import { ReplayBrowserPage } from "./replays_page";
+
+export function createReplaysPage(): { Page: React.ComponentType } {
+  const Page = () => <ReplayBrowserPage />;
+
+  return { Page };
+}
