+++ src/renderer/pages/replays/replay_browser/replay_file/stories/stage.png
