+++ src/renderer/pages/replays/replay_browser/replay_file/team_elements/player_badge/player_badge.stories.tsx
@@ -1,4 +1,4 @@
-import { PlayerBadge } from "./PlayerBadge";
+import { PlayerBadge } from "./player_badge";
 
 type PlayerBadgeProps = React.ComponentProps<typeof PlayerBadge>;
