+++ src/renderer/pages/replays/replay_browser/replay_file/team_elements/player_badge/player_badge.tsx
