+++ src/renderer/pages/replays/replay_file_stats/home_run_profile.tsx
