+++ src/renderer/pages/replays/replay_file_stats/table_styles.tsx
